requirements.txt
opencv-python
matplotlib
numpy
pillow
imageio
pandas
tqdm